Day 18, part 2

+++ src/test/kotlin/io/github/davidmerrick/aoc2020/day18/Day18Test.kt
@@ -69,7 +69,7 @@ class Day18Test {
     fun `Part 1`() {
         val result = TestUtil.parseInput(this::class, "day18.txt") {
             val postFix = ExpressionEvaluator.toPostFix(parse(it))
-            ExpressionEvaluator.evaluatePostFix(postFix).toLong()
+            ExpressionEvaluator.evaluatePostFix(postFix)
         }.sum()
         println(result)
     }
